upstream: add rook operator
Create a new operator under community-operators for Rook the Storage
Orchestrator for Kubernetes. This is the Kubernetes upstream version.

The initial implementation expects the Ceph cluster to be part of the
same namespace as the operator.
It also has all Rook's capabilities when it comes to creating, managing
and upgrading a cluster. Simply edit the cluster CR to apply any changes
to your deployment.

Signed-off-by: Sébastien Han <seb@redhat.com>