Use operator-registry format for upstream
Adjust the upstream community operators to use the newer
nested operator-registry package format for all operators in the
upstream-community-operators directory. This is being done to remove the
need to call `operator-courier nest` against all of the operators when
the catalog is built by the `upstream.Dockerfile`. This commit also
updates the dockerfile to remove that build step.