Submariner: fix CRD validation
The current Submariner CRD defines validation criteria inside the
per-version stanza, which is no longer allowed when all the criteria
are identical for all published versions (which is the case here,
since there's only one version).

We'd fixed this for Kubernetes 1.15 compatibility, but hadn't updated
the resources here; this patch updates the CRD as appropriate.

Signed-off-by: Stephen Kitt <skitt@redhat.com>