Change labeling on spark operator deployments
If labels on the operator deployment change, automatic updates
through OLM do not work for running deployments (selectors and
match expressions are immutable).  Drop the version label, and
change the name to be distinct from the gcp spark operator.

Automatic updates from 1.0.8 to 1.1.0 will still fail for running operators,
but updates will work in the future.

Signed-off-by: Trevor McKay <tmckay@redhat.com>